Game technically done
Made it so the tide can go up. have it spawn enemies and items when tide is going down. Made it when you die you teleport to the starting location, player start. Added a print string and check for getting 3 progress to win game. Changed enemy to be the crab model and added anim playing. Various fixes to the player, enemy, and items.